chore(demo): move KeyPairs to crypto namespace
